betmaker/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.78.0
frozenlist==1.3.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
multidict==6.0.2
pydantic==1.9.1
pypika-tortoise==0.1.5
pytz==2022.1
sniffio==1.2.0
starlette==0.19.1
tortoise-orm==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
yarl==1.7.2
